Improve handling of missing ConfirmProvider
Prior to 3.0.13, material-ui-confirm allowed usage without a
`ConfirmProvider`, as long as it wasn't invoked.  This was useful for test
purposes; automated tests of components that happened to call
`useConfirm` could skip rendering `ConfirmProvider` as long as they
didn't actually depend on showing a confirmation prompt.

Starting with 3.0.13, code instead throws errors similar to the
following:

```
    Missing ConfirmProvider

      55 |       await Promise.race([
      56 |         exited,
    > 57 |         confirm(useOptions)
         |         ^
      58 |           .then(() => (confirmed = true))
      59 |           .catch(() => (confirmed = false)),
      60 |       ]);

      at confirmBase (../../node_modules/material-ui-confirm/dist/material-ui-confirm.cjs.js:144:11)
      at ../../node_modules/material-ui-confirm/dist/material-ui-confirm.cjs.js:350:12
```

This PR adds a default `ConfirmProvider` context that does nothing if
confirmation prompts are not shown and throws a clear error message if a
confirmation prompt is attempted.