Prevent flash of defaults during close (#99)
Using "state == null" to indicate "not open" means that during the fade-out
transition the options all revert to defaults, so you see a brief flash of the
default title/content.

Switched to using an explicit 'open' flag within state, so the options are
preserved during the close transition. State will still be null before the first
dialog is opened, but otherwise will retain the last options used.